added command to manage channel groups added ability to disable command in channel groups refactored embed handling to support multiple embeds to be send at once (handling to decide to split into multiple embeds is still needed)

+++ abstracto-application/core/core-impl/src/main/java/dev/sheldan/abstracto/core/service/ChannelGroupServiceBean.java
@@ -1,9 +1,16 @@
 package dev.sheldan.abstracto.core.service;
 
+import dev.sheldan.abstracto.core.command.exception.ChannelGroupException;
+import dev.sheldan.abstracto.core.command.exception.CommandException;
+import dev.sheldan.abstracto.core.command.models.ACommand;
+import dev.sheldan.abstracto.core.command.service.management.ChannelGroupCommandManagementService;
+import dev.sheldan.abstracto.core.command.service.management.CommandManagementService;
 import dev.sheldan.abstracto.core.models.database.AChannel;
 import dev.sheldan.abstracto.core.models.database.AChannelGroup;
+import dev.sheldan.abstracto.core.models.database.AServer;
 import dev.sheldan.abstracto.core.service.management.ChannelGroupManagementService;
 import dev.sheldan.abstracto.core.service.management.ChannelManagementService;
+import dev.sheldan.abstracto.core.service.management.ServerManagementService;
 import net.dv8tion.jda.api.entities.TextChannel;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -17,9 +24,25 @@ public class ChannelGroupServiceBean implements ChannelGroupService {
     @Autowired
     private ChannelManagementService channelManagementService;
 
+    @Autowired
+    private CommandManagementService commandManagementService;
+
+    @Autowired
+    private ChannelGroupCommandManagementService channelGroupCommandManagementService;
+
+    @Autowired
+    private ServerManagementService serverManagementService;
+
     @Override
-    public AChannelGroup createChannelGroup(String name) {
-        return channelGroupManagementService.createChannelGroup(name);
+    public AChannelGroup createChannelGroup(String name, Long serverId) {
+        AServer server = serverManagementService.loadOrCreate(serverId);
+        return channelGroupManagementService.createChannelGroup(name, server);
+    }
+
+    @Override
+    public void deleteChannelGroup(String name, Long serverId) {
+        AServer server = serverManagementService.loadOrCreate(serverId);
+        channelGroupManagementService.deleteChannelGroup(name, server);
     }
 
     @Override
@@ -35,7 +58,11 @@ public class ChannelGroupServiceBean implements ChannelGroupService {
 
     @Override
     public void addChannelToChannelGroup(String channelGroupName, AChannel channel) {
-        AChannelGroup channelGroup = channelGroupManagementService.findByName(channelGroupName);
+        AServer server = serverManagementService.loadOrCreate(channel.getServer().getId());
+        AChannelGroup channelGroup = channelGroupManagementService.findByNameAndServer(channelGroupName, server);
+        if(channelGroup == null) {
+            throw new ChannelGroupException(String.format("Channel group %s was not found.", channelGroupName));
+        }
         channelGroupManagementService.addChannelToChannelGroup(channelGroup, channel);
     }
 
@@ -52,7 +79,45 @@ public class ChannelGroupServiceBean implements ChannelGroupService {
 
     @Override
     public void removeChannelFromChannelGroup(String channelGroupName, AChannel channel) {
-        AChannelGroup channelGroup = channelGroupManagementService.findByName(channelGroupName);
+        AServer server = serverManagementService.loadOrCreate(channel.getServer().getId());
+        AChannelGroup channelGroup = channelGroupManagementService.findByNameAndServer(channelGroupName, server);
+        if(channelGroup == null) {
+            throw new ChannelGroupException(String.format("Channel group %s was not found", channelGroupName));
+        }
         channelGroupManagementService.removeChannelFromChannelGroup(channelGroup, channel);
+    }
+
+    @Override
+    public void disableCommandInChannelGroup(String commandName, String channelGroupName, Long serverId) {
+        AServer server = serverManagementService.loadOrCreate(serverId);
+        AChannelGroup channelGroup = channelGroupManagementService.findByNameAndServer(channelGroupName, server);
+        if(channelGroup == null) {
+            throw new ChannelGroupException(String.format("Channel group %s was not found", channelGroupName));
+        }
+        ACommand command = commandManagementService.findCommandByName(commandName);
+        if(command == null) {
+            throw new CommandException(String.format("Command %s not found.", commandName));
+        }
+        channelGroupCommandManagementService.setCommandInGroupTo(command, channelGroup, false);
+    }
+
+    @Override
+    public void enableCommandInChannelGroup(String commandName, String channelGroupName, Long serverId) {
+        AServer server = serverManagementService.loadOrCreate(serverId);
+        AChannelGroup channelGroup = channelGroupManagementService.findByNameAndServer(channelGroupName, server);
+        if(channelGroup == null) {
+            throw new ChannelGroupException(String.format("Channel group %s was not found", channelGroupName));
+        }
+        ACommand command = commandManagementService.findCommandByName(commandName);
+        if(command == null) {
+            throw new CommandException(String.format("Command %s not found.", commandName));
+        }
+        channelGroupCommandManagementService.setCommandInGroupTo(command, channelGroup, true);
+    }
+
+    @Override
+    public boolean doesGroupExist(String groupName, Long serverId) {
+        AServer server = serverManagementService.loadOrCreate(serverId);
+        return channelGroupManagementService.findByNameAndServer(groupName, server) != null;
     }
 }
